+++ MODULE_GLOB.mjs
@@ -2,7 +2,9 @@
 
 /**
  * File glob pattern to recursively match TypeScript (`.mts`, `.cts`, `.ts`, and
- * `.tsx`; excluding TypeScript definition files) and JavaScript (`.mjs`,
- * `.cjs`, `.js`, and `.jsx`) modules.
+ * `.tsx`) and JavaScript (`.mjs`, `.cjs`, `.js`, and `.jsx`) modules, including
+ * dotfiles.
  */
-export default "**/{!(*.d).mts,!(*.d).cts,!(*.d).ts,*.{mjs,cjs,js,jsx,tsx}}";
+const MODULE_GLOB = `**/{,*,.*}.{mts,cts,ts,tsx,mjs,cjs,js,jsx}`;
+
+export default MODULE_GLOB;