+++ test/fixtures/side-effect-import/a.mjs
@@ -1,0 +1,3 @@
+import "./a.mjs";
+
+export default 1;