+++ test/fixtures/unresolvable-import-specifier/a.mjs
@@ -1,2 +1,4 @@
 // eslint-disable-next-line no-unused-vars
 import a from "./b.mjs";
+
+export default 1;